# Minor performance tweaks from profiling