--- a/src/Bicep.Core/Semantics/Namespaces/AzNamespaceType.cs
+++ b/src/Bicep.Core/Semantics/Namespaces/AzNamespaceType.cs
@@ -36,6 +36,11 @@ namespace Bicep.Core.Semantics.Namespaces
             ArmTemplateProviderName: "AzureResourceManager",
             ArmTemplateProviderVersion: new Version(EmbeddedAzProviderVersion).ToString(3));
 
+        private delegate bool VisibilityDelegate(ResourceScope scope, BicepSourceFileKind sourceFileKind);
+        private record NamespaceValue<T>(T Value, VisibilityDelegate IsVisible);
+
+        private static readonly ImmutableArray<NamespaceValue<FunctionOverload>> Overloads = GetAzOverloads().ToImmutableArray();
+
         private static FunctionOverload.ResultBuilderDelegate AddDiagnosticsAndReturnResult(TypeSymbol returnType, DiagnosticBuilder.DiagnosticBuilderDelegate writeDiagnostic)
         {
             return (_, _, _, diagnostics, functionCall, argumentTypes) =>
@@ -188,7 +193,7 @@ namespace Bicep.Core.Semantics.Namespaces
             }, null);
         }
 
-        private static ObjectType GetDeploymentReturnType(ResourceScope targetScope)
+        private static ObjectType GetDeploymentReturnType(bool resourceGroupScope)
         {
             // Note: there are other properties which could be included here, but they allow you to break out of the bicep world.
             // We're going to omit them and only include what is truly necessary. If we get feature requests to expose more properties, we should discuss this further.
@@ -211,7 +216,7 @@ namespace Bicep.Core.Semantics.Namespaces
                 }, null)),
             };
 
-            if (!targetScope.HasFlag(ResourceScope.ResourceGroup))
+            if (!resourceGroupScope)
             {
                 // deployments in the 'resourcegroup' scope do not have the 'location' property. All other scopes do.
                 var locationProperty = new TypeProperty("location", LanguageConstants.String);
@@ -295,11 +300,25 @@ namespace Bicep.Core.Semantics.Namespaces
                     .WithRequiredParameter("resourceGroupName", LanguageConstants.String, "The resource group name")
                     .Build(),
                 ResourceScope.Tenant | ResourceScope.ManagementGroup | ResourceScope.Subscription | ResourceScope.ResourceGroup);
+
+            yield return (
+                new FunctionOverloadBuilder("deployment")
+                    .WithReturnType(GetDeploymentReturnType(resourceGroupScope: true))
+                    .WithGenericDescription("Returns information about the current deployment operation.")
+                    .Build(),
+                ResourceScope.ResourceGroup);
+
+            yield return (
+                new FunctionOverloadBuilder("deployment")
+                    .WithReturnType(GetDeploymentReturnType(resourceGroupScope: false))
+                    .WithGenericDescription("Returns information about the current deployment operation.")
+                    .Build(),
+                ResourceScope.Tenant | ResourceScope.ManagementGroup | ResourceScope.Subscription);
         }
 
-        private static IEnumerable<FunctionOverload> GetAzOverloads(ResourceScope resourceScope, BicepSourceFileKind sourceFileKind)
+        private static IEnumerable<NamespaceValue<FunctionOverload>> GetAzOverloads()
         {
-            if (sourceFileKind == BicepSourceFileKind.ParamsFile)
+            static IEnumerable<FunctionOverload> GetParamsFilePermittedOverloads()
             {
                 yield return new FunctionOverloadBuilder(GetSecretFunctionName)
                     .WithReturnType(LanguageConstants.SecureString)
@@ -344,25 +363,9 @@ namespace Bicep.Core.Semantics.Namespaces
                     .WithFlags(FunctionFlags.DirectAssignment)
                     .Build();
             }
-            else
+
+            static IEnumerable<FunctionOverload> GetBicepFilePermittedOverloads()
             {
-                foreach (var (functionOverload, allowedScopes) in GetScopeFunctions())
-                {
-                    // we only include it if it's valid at all of the scopes that the template is valid at
-                    if (resourceScope == (resourceScope & allowedScopes))
-                    {
-                        yield return functionOverload;
-                    }
-
-                    // TODO: add banned function to explain why a given function isn't available
-                }
-
-                // TODO: Add schema for return type
-                yield return new FunctionOverloadBuilder("deployment")
-                    .WithReturnType(GetDeploymentReturnType(resourceScope))
-                    .WithGenericDescription("Returns information about the current deployment operation.")
-                    .Build();
-
                 yield return new FunctionOverloadBuilder("environment")
                     .WithReturnType(GetEnvironmentReturnType())
                     .WithGenericDescription("Returns information about the Azure environment used for deployment.")
@@ -500,9 +503,25 @@ namespace Bicep.Core.Semantics.Namespaces
                     .WithFlags(FunctionFlags.RequiresInlining)
                     .Build();
             }
+
+            foreach (var overload in GetBicepFilePermittedOverloads())
+            {
+                yield return new(overload, (_, sfk) => sfk == BicepSourceFileKind.BicepFile);
+            }
+
+            foreach (var overload in GetParamsFilePermittedOverloads())
+            {
+                yield return new(overload, (_, sfk) => sfk == BicepSourceFileKind.ParamsFile);
+            }
+
+            foreach (var (overload, allowedScopes) in GetScopeFunctions())
+            {
+                // we only include it if it's valid at all of the scopes that the template is valid at
+                yield return new(overload, (scope, sfk) => sfk == BicepSourceFileKind.BicepFile && scope == (scope & allowedScopes));
+            }
         }
 
-        public static NamespaceType Create(string aliasName, ResourceScope resourceScope, IResourceTypeProvider resourceTypeProvider, BicepSourceFileKind bicepSourceFileKind)
+        public static NamespaceType Create(string aliasName, ResourceScope scope, IResourceTypeProvider resourceTypeProvider, BicepSourceFileKind sourceFileKind)
         {
             return new NamespaceType(
                 aliasName,
@@ -513,7 +532,7 @@ namespace Bicep.Core.Semantics.Namespaces
                     ArmTemplateProviderName: "AzureResourceManager",
                     ArmTemplateProviderVersion: resourceTypeProvider.Version),
                 ImmutableArray<TypeProperty>.Empty,
-                GetAzOverloads(resourceScope, bicepSourceFileKind),
+                Overloads.Where(x => x.IsVisible(scope, sourceFileKind)).Select(x => x.Value),
                 ImmutableArray<BannedFunction>.Empty,
                 ImmutableArray<Decorator>.Empty,
                 resourceTypeProvider);