--- a/src/Bicep.Core/TypeSystem/DecoratorResolver.cs
+++ b/src/Bicep.Core/TypeSystem/DecoratorResolver.cs
@@ -24,7 +24,7 @@ namespace Bicep.Core.TypeSystem
 
         public Decorator? TryGetDecorator(FunctionOverload overload) => this.decoratorsByOverloads.TryGetValue(overload, out Decorator? decorator) ? decorator : null;
 
-        public ImmutableDictionary<string, FunctionSymbol> GetKnownDecoratorFunctions() => this.functionResolver.GetKnownFunctions();
+        public IReadOnlyDictionary<string, FunctionSymbol> GetKnownDecoratorFunctions() => this.functionResolver.GetKnownFunctions();
 
         public IEnumerable<Decorator> GetMatches(FunctionSymbol symbol, IList<TypeSymbol> argumentTypes)
         {