--- a/src/Bicep.Tools.Benchmark/Compilation.cs
+++ b/src/Bicep.Tools.Benchmark/Compilation.cs
@@ -26,7 +26,7 @@ public class Compilation
 
     private static BenchmarkData CreateBenchmarkData()
     {
-        var fileSystem = FileHelper.CreateMockFileSystemForEmbeddedFiles(typeof(DataSet).Assembly, "Files");
+        var fileSystem = FileHelper.CreateMockFileSystemForEmbeddedFiles(typeof(DataSet).Assembly, "Files/baselines");
 
         var dataSets = DataSets.AllDataSets
             .Where(x => !x.HasRegistryModules)