… src/Bicep.Core/Semantics/Namespaces/SystemNamespaceType.cs