Tested by running `dotnet run --configuration Release --project src/Bicep.Tools.Benchmark -- --filter '*Compilation*' --profiler EP`

Results from main branch:
```
| Method                                         | Mean     | Error    | StdDev   | Gen0      | Gen1      | Allocated |
|----------------------------------------------- |---------:|---------:|---------:|----------:|----------:|----------:|
| 'Get compilation diagnostics for each dataset' | 989.7 ms | 19.53 ms | 49.00 ms | 3000.0000 | 2000.0000 | 285.98 MB |
```

Results from this branch:
```
| Method                                         | Mean     | Error    | StdDev   | Gen0      | Gen1      | Allocated |
|----------------------------------------------- |---------:|---------:|---------:|----------:|----------:|----------:|
| 'Get compilation diagnostics for each dataset' | 815.2 ms | 17.37 ms | 50.93 ms | 3000.0000 | 2000.0000 | 257.16 MB |
```
 ###### Microsoft Reviewers: [Open in CodeFlow](https://microsoft.github.io/open-pr/?codeflow=https://github.com/Azure/bicep/pull/12994)