@StephenWeatherford 2024-01-17
So what exactly was the issue?  Immutable collections are used quite extensively, do we need to avoid them?