@anthony-c-martin 2024-01-18
> So what exactly was the issue? Immutable collections are used quite extensively, do we need to avoid them?
It's context dependent - what I found is that short-lived **large** immutable collections are pretty expensive. The impression I got was that immutable collections are more optimized for long-lived collections with safe mutations. The syntax hierarchy is rebuilt from scratch on every keypress, so we're not really getting the benefits.

The other optimization was just to avoid re-creating all of the namespace function definitions on every keypress.